pyproject.toml
[tool.poetry]
name = "meta"
version = "0.0.1"
description = "Prism Launcher meta generator"
authors = ["Sefa Eyeoglu <contact@scrumplex.net>"]
license = "MS-PL"
readme = "README.md"
repository = "https://github.com/PrismLauncher/meta"
documentation = "https://github.com/PrismLauncher/meta"
keywords = ["metadata", "prism", "launcher"]

[tool.poetry.scripts]
generateFabric = "meta.run.generate_fabric:main"
generateForge = "meta.run.generate_forge:main"
generateLiteloader = "meta.run.generate_liteloader:main"
generateMojang = "meta.run.generate_mojang:main"
generateNeoForge = "meta.run.generate_neoforge:main"
generateQuilt = "meta.run.generate_quilt:main"
generateJava = "meta.run.generate_java:main"
generateAuthlibInjector = "meta.run.generate_authlib_injector:main"
updateFabric = "meta.run.update_fabric:main"
updateForge = "meta.run.update_forge:main"
updateLiteloader = "meta.run.update_liteloader:main"
updateMojang = "meta.run.update_mojang:main"
updateNeoForge = "meta.run.update_neoforge:main"
updateQuilt = "meta.run.update_quilt:main"
updateJava = "meta.run.update_java:main"
updateAuthlibInjector = "meta.run.update_authlib_injector:main"
index = "meta.run.index:main"

[tool.poetry.dependencies]
python = "^3.8"
cachecontrol = "^0.14.0"
requests = "^2.31.0"
filelock = "^3.13.1"
packaging = "^25.0"
pydantic = "^1.10.13"
python-dateutil = "^2.9.0.post0"
pygithub = "^2.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
